backend/requirements.txt
ruff
mypy
SQLAlchemy
alembic
fastapi
httpx
itsdangerous
pydantic
pydantic-settings
uvicorn
jinja2
factory-boy
types-factory-boy
pytest
pytest-xdist
pytest-asyncio
pytest-factoryboy
colorlog
